# orchestrator/server/docs/swagger.yaml
basePath: /api
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  handlers.CreateWorkflowRequest:
    properties:
      description:
        description: 工作流描述
        example: 单步监控流水线
        type: string
      name:
        description: 工作流名称
        example: monitor
        type: string
      sharedContext:
        additionalProperties: {}
        description: 跨步骤共享上下文（步骤间传参）
        type: object
      steps:
        description: '步骤列表（type: script/wait/condition/screenshot；通过 dependsOn 声明依赖）'
        items:
          $ref: '#/definitions/models.WorkflowStep'
        type: array
    required:
    - name
    - steps
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        description: Error 人类可读的错误说明
        example: insufficient permissions
        type: string
      success:
        description: Success 固定为 false
        example: false
        type: boolean
    type: object
...
  handlers.ListWorkflowsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Workflow'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
  handlers.StepStatusResponse:
    properties:
      data:
        $ref: '#/definitions/models.StepStatus'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.SuccessResponse:
    properties:
      success:
        example: true
        type: boolean
    type: object
...
  handlers.WorkerEntry:
    properties:
      endTime:
        description: 结束时间（毫秒）
        example: 1712937620000
        type: integer
      startTime:
        description: 开始时间（毫秒）
        example: 1712937610000
        type: integer
      status:
        description: 步骤状态
        example: completed
        type: string
      stepId:
        description: 该 worker 承担的首个步骤 ID
        example: s1
        type: string
      workerId:
        description: Agent ID
        example: agent-001
        type: string
    type: object
  handlers.WorkerListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.WorkerEntry'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.WorkflowDetailData:
    properties:
      createdTime:
        description: 创建时间（毫秒）
        example: 1712937600000
        type: integer
      description:
        description: 描述
        example: 单步监控流水线
        type: string
      endTime:
        description: 结束时间（毫秒，未结束为 0）
        example: 1712937620000
        type: integer
      id:
        description: 工作流 ID
        example: 1
        type: integer
      name:
        description: 名称
        example: monitor
        type: string
      sharedContext:
        additionalProperties: {}
        description: 共享上下文快照
        type: object
      startTime:
        description: 开始时间（毫秒，未开始为 0）
        example: 1712937610000
        type: integer
      status:
        description: 状态：pending/running/completed/failed/cancelled
        example: completed
        type: string
      stepStatus:
        additionalProperties: {}
        description: 各步骤实时状态（stepId → 状态）
        type: object
      steps:
        description: 步骤定义列表
        items:
          $ref: '#/definitions/models.WorkflowStep'
        type: array
    type: object
  handlers.WorkflowDetailResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.WorkflowDetailData'
      success:
        example: true
        type: boolean
    type: object
...
  models.StepStatus:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      endTime:
        type: string
      id:
        type: integer
      message:
        type: string
      name:
        type: string
      startTime:
        type: string
      status:
        description: pending/running/completed/failed/skipped
        type: string
      stepId:
        type: string
      updatedAt:
        type: string
      workerId:
        type: string
      workflowId:
        type: integer
    type: object
  models.Workflow:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      endTime:
        type: string
      id:
        type: integer
      name:
        type: string
      startTime:
        type: string
      status:
        description: pending/running/completed/failed/cancelled
        type: string
      updatedAt:
        type: string
    type: object
  models.WorkflowStep:
    properties:
      dependsOn:
        items:
          type: string
        type: array
      id:
        type: string
      maxRetries:
        description: 失败重试次数（0=不重试）
        type: integer
      name:
        type: string
      parameters:
        additionalProperties: true
        type: object
      retryBackoffSeconds:
        description: 重试退避基数（秒，默认 2，指数翻倍）
        type: integer
      script:
        type: string
      timeoutSeconds:
        type: integer
      type:
        description: 步骤类型：script(默认)/wait/condition/screenshot
        type: string
      workers:
        items:
          type: string
        type: array
    type: object
host: localhost:9527
info:
  contact: {}
  description: |-
    Wingman 远程中控编排器 HTTP API（auth / RBAC / agent / script / workflow / messages / settings）。
    架构约束：Dashboard 与远程客户端只连接本 server，不直连 runtime。
  title: Wingman Orchestrator API
  version: "1.0"
paths:
...
  /workflow-templates:
    get:
      description: 内置模板目录：单步监控/并行采集/串行流水线/fan-out/独立步骤类型 wait-condition-screenshot；可直接作为创建工作流的
        steps 骨架
      produces:
      - application/json
      responses:
        "200":
          description: data 为模板数组（name/description/steps）
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 工作流模板库
      tags:
      - workflows
  /workflows:
    get:
      description: 返回所有工作流记录（含状态：pending/running/completed/failed/cancelled）；任何登录用户可读
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListWorkflowsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: 数据库查询失败
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 工作流列表
      tags:
      - workflows
    post:
      consumes:
      - application/json
      description: 校验步骤定义（依赖关系/worker 存在性）后提交引擎调度，调度失败时回滚 DB 记录；需要 workflows:run 权限
      parameters:
      - description: 工作流定义
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWorkflowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: data.workflowId 为新建工作流 ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 步骤校验失败或请求体格式错误
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: 序列化或 DB 写入失败
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建工作流
      tags:
      - workflows
  /workflows/{id}:
    get:
      description: 返回工作流定义、步骤状态与共享上下文快照（运行中与已结束均可查询）
      parameters:
      - description: 工作流 ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkflowDetailResponse'
        "400":
          description: id 非数字
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: 工作流不存在
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 工作流详情
      tags:
      - workflows
  /workflows/{id}/cancel:
    post:
      description: 取消运行中的工作流（通知引擎中止后续步骤）并写审计日志；需要 workflows:run 权限
      parameters:
      - description: 工作流 ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: id 非数字或工作流不可取消
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取消工作流
      tags:
      - workflows
  /workflows/{id}/steps/{stepId}/status:
    get:
      description: 合并 DB 记录与引擎内存快照，返回单步的实时执行状态（status/message/workerId/起止时间）；任何登录用户可读
      parameters:
      - description: 工作流 ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: 步骤 ID
        example: s1
        in: path
        name: stepId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StepStatusResponse'
        "400":
          description: id 非数字
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: 步骤状态记录不存在
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 工作流步骤状态
      tags:
      - workflows
  /workflows/{id}/workers:
    get:
      description: 汇总步骤状态中按 workerId 去重的 worker（agent）信息；任何登录用户可读
      parameters:
      - description: 工作流 ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkerListResponse'
        "400":
          description: id 非数字
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 工作流 worker 列表
      tags:
      - workflows
securityDefinitions:
  BearerAuth:
    description: '"Bearer <JWT>"（由 POST /api/v1/auth/login 获取）'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"